requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
factory-boy==3.3.0
Faker==24.4.0
inflection==0.5.1
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
pytest==8.1.1
pytest-django==4.8.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
